Switch from OPA to Casl (#78)
* Enable ts

* Fixes: dirname

* Fixed: jsdoc return type

* Implemented OPA functionality in Casl

* Tests for Casl

* Swapped OPA out for Casl

* Fix: Test import

* Formated Casl tests

* Properly filter machines + imported missing type

* Support for role expiration for Abilities

* Removed unused function

* Fixed authorization constraints for roles

* Case-insensitive comparison in Casl rules

* removed --openssl-legacy-provider for ms

* lint pass

* Replaced lru-cache for redis

* Linter pass

---------

+++ src/management-system/src/backend/server/rest-api/process.js
@@ -15,21 +15,10 @@ import {
   getProcessImage,
 } from '../../shared-electron-server/data/process.js';
 import express from 'express';
-import { isAllowed } from '../iam/middleware/authorization.js';
-import {
-  PERMISSION_VIEW,
-  PERMISSION_CREATE,
-  PERMISSION_UPDATE,
-  PERMISSION_DELETE,
-  PERMISSION_MANAGE,
-} from '../../../shared-frontend-backend/constants/index.js';
-import { config } from '../iam/utils/config.js';
-import { ensureOpaSync } from '../iam/opa/opa-client.js';
+import { isAllowed } from '../iam/middleware/authorization.ts';
 import logger from '../../shared-electron-server/logging.js';
-import bpmnHelper from '@proceed/bpmn-helper';
-const { addDocumentation, setDefinitionsName } = bpmnHelper;
-import processHelpers from '../../../shared-frontend-backend/helpers/processHelpers.js';
-const { getProcessInfo } = processHelpers;
+import Ability from '../iam/authorization/abilityHelper';
+import { toCaslResource } from '../iam/authorization/caslRules';
 
 const processRouter = express.Router();
 
@@ -42,36 +31,31 @@ function toExternalFormat(processMetaData) {
   return newFormat;
 }
 
-processRouter.use(
-  '/',
-  isAllowed(PERMISSION_VIEW, 'Process', { filter: true }),
-  async (req, res, next) => {
-    if (config.useAuthorization) {
-      req.processes = req.filter.map((process) => toExternalFormat(process));
-    } else {
-      req.processes = getProcesses().map((process) => toExternalFormat(process));
-    }
-    next();
-  },
-);
-
-processRouter.get('/', async (req, res) => {
+processRouter.get('/', isAllowed('view', 'Process'), async (req, res) => {
   let { noBpmn } = req.query;
 
+  const processes = getProcesses();
+
+  /** @type {Ability} */
+  const userAbility = req.userAbility;
+
   try {
-    const processes = await Promise.all(
-      req.processes.map(async (process) => {
-        if (noBpmn === 'true') {
-          return process;
-        } else {
-          const bpmn = await getProcessBpmn(process.definitionId);
-          return { ...process, bpmn };
-        }
-      }),
+    const userProcessesPromise = await Promise.allSettled(
+      userAbility
+        .filter('view', 'Process', processes)
+        .map(async (process) =>
+          noBpmn === 'true'
+            ? process
+            : { ...process, bpmn: await getProcessBpmn(process.definitionId) },
+        ),
     );
 
-    if (processes.length) {
-      return res.status(200).json(processes);
+    const userProcesses = userProcessesPromise
+      .filter((promise) => promise.status === 'fulfilled')
+      .map((promise) => toExternalFormat(promise.value));
+
+    if (userProcesses.length) {
+      return res.status(200).json(userProcesses);
     } else {
       return res.status(204).end();
     }
@@ -80,51 +64,59 @@ processRouter.get('/', async (req, res) => {
   }
 });
 
-processRouter.post(
-  '/',
-  isAllowed([PERMISSION_CREATE, PERMISSION_MANAGE], 'Process'),
-  async (req, res) => {
-    const { body } = req;
+processRouter.post('/', isAllowed('create', 'Process'), async (req, res) => {
+  const { body } = req;
 
-    if (typeof body !== 'object') {
-      res.status(400).send('Expected { bpmn: ... } as body!');
-      return;
+  if (typeof body !== 'object') {
+    res.status(400).send('Expected { bpmn: ... } as body!');
+    return;
+  }
+
+  /** @type {Ability} */
+  const userAbility = req.userAbility;
+
+  if (!userAbility.can('create', toCaslResource('Process', body)))
+    return res.status(403).send('Forbidden.');
+
+  const { bpmn } = body;
+
+  try {
+    const process = await addProcess(body);
+    if (typeof process === 'object') {
+      res.status(201).json(toExternalFormat({ ...process, bpmn }));
+    } else {
+      // a process with this id does already exist
+      res.status(303).location(`/process/${process}`).send();
     }
 
-    const { bpmn } = body;
-
-    try {
-      const process = await addProcess(body);
-      if (typeof process === 'object') {
-        res.status(201).json(toExternalFormat({ ...process, bpmn }));
-        await ensureOpaSync(`processes/${process.id}`, undefined, process);
-      } else {
-        // a process with this id does already exist
-        res.status(303).location(`/process/${process}`).send();
-      }
-
-      return;
-    } catch (err) {
-      logger.debug(`Error on POST request to /process: ${err}`);
-      res.status(500).send('Failed to create the process due an internal error');
-    }
-  },
-);
+    return;
+  } catch (err) {
+    logger.debug(`Error on POST request to /process: ${err}`);
+    res.status(500).send('Failed to create the process due an internal error');
+  }
+});
 
 processRouter.use('/:definitionId', async (req, res, next) => {
   req.definitionsId = req.params.definitionId;
-  // use processes from earlier middleware to allow filtering in one place
-  req.process = req.processes.find((p) => p.definitionId === req.definitionsId);
+  req.process = getProcesses().find((process) => {
+    return process.id === req.params.definitionId;
+  });
   next();
 });
 
-processRouter.get('/:definitionId', async (req, res) => {
+processRouter.get('/:definitionId', isAllowed('view', 'Process'), async (req, res) => {
   const { process, definitionsId } = req;
 
   if (!process) {
     res.status(404).send(`Process with id ${definitionsId} could not be found!`);
     return;
   }
+
+  /** @type {Ability} */
+  const userAbility = req.userAbility;
+
+  if (!userAbility.can('view', toCaslResource('Process', process)))
+    return res.status(403).send('Forbidden.');
 
   try {
     const bpmn = await getProcessBpmn(definitionsId);
@@ -134,137 +126,114 @@ processRouter.get('/:definitionId', async (req, res) => {
   }
 });
 
-const allowedInPutBodyInsteadOfBpmn = ['definitionName', 'description'];
-processRouter.put(
-  '/:definitionId',
-  isAllowed([PERMISSION_UPDATE, PERMISSION_MANAGE], 'Process'),
-  async (req, res) => {
-    const { process, definitionsId, body } = req;
+processRouter.put('/:definitionId', isAllowed('update', 'Process'), async (req, res) => {
+  const { process, definitionsId, body } = req;
 
-    if (!process) {
-      res.status(404).send('Process to update does not exist!');
-      return;
-    }
-
-    if (typeof body !== 'object') {
-      res.status(400).send('Expected { bpmn: ... } as body!');
-      return;
-    }
-
-    let { bpmn } = body;
-
-    if (
-      typeof bpmn === 'string' &&
-      bpmn !== '' &&
-      allowedInPutBodyInsteadOfBpmn.some((key) => Object.keys(body).includes(key))
-    ) {
-      const oldBpmn = await getProcessBpmn(definitionsId);
-      const dataInOldBpmn = await getProcessInfo(oldBpmn);
-      const dataInNewBpmn = await getProcessInfo(bpmn);
-
-      if ('definitionName' in body) {
-        if (dataInNewBpmn.name !== dataInOldBpmn.name)
-          return res
-            .status(400)
-            .send(
-              'Key "definitionName" of process being updated in the request body and in the BPMN.',
-            );
-
-        bpmn = await setDefinitionsName(bpmn, body['definitionName']);
-        delete body['definitionName'];
-      }
-
-      if ('description' in body) {
-        if (dataInNewBpmn.description !== dataInOldBpmn.description)
-          return res
-            .status(400)
-            .send(
-              'Key "description" of process being updated in the request body and in the BPMN.',
-            );
-
-        bpmn = await addDocumentation(bpmn, body['description']);
-        delete body['description'];
-      }
-
-      body.bpmn = bpmn;
-    }
-
-    try {
-      const newProcessInfo = await updateProcess(definitionsId, body);
-      bpmn = bpmn || (await getProcessBpmn(definitionsId));
-      res.status(200).json(toExternalFormat({ ...newProcessInfo, bpmn }));
-      await ensureOpaSync(`processes/${definitionsId}`, undefined, newProcessInfo);
-      return;
-    } catch (err) {
-      res.status(400).send(err.message);
-    }
-  },
-);
-
-processRouter.delete(
-  '/:definitionId',
-  isAllowed([PERMISSION_DELETE, PERMISSION_MANAGE], 'Process'),
-  async (req, res) => {
-    const { process, definitionsId } = req;
-
-    if (process) {
-      await removeProcess(definitionsId);
-    }
-
-    res.status(200).end();
-    await ensureOpaSync(`processes/${definitionsId}`, 'DELETE');
-    return;
-  },
-);
-
-processRouter.get('/:definitionId/versions', async (req, res) => {
-  const { process, definitionsId } = req;
   if (!process) {
-    res.status(404).send(`Process with id ${definitionsId} could not be found!`);
+    res.status(404).send('Process to update does not exist!');
     return;
   }
+
+  if (typeof body !== 'object') {
+    res.status(400).send('Expected { bpmn: ... } as body!');
+    return;
+  }
+
+  /** @type {Ability} */
+  const userAbility = req.userAbility;
+
+  if (!userAbility.can('update', toCaslResource('Process', process)))
+    return res.status(403).send('Forbidden.');
+
+  let { bpmn } = body;
+
+  try {
+    const newProcessInfo = await updateProcess(definitionsId, body);
+    bpmn = bpmn || (await getProcessBpmn(definitionsId));
+    res.status(200).json(toExternalFormat({ ...newProcessInfo, bpmn }));
+    return;
+  } catch (err) {
+    res.status(400).send(err.message);
+  }
+});
+
+processRouter.delete('/:definitionId', isAllowed('delete', 'Process'), async (req, res) => {
+  const { process, definitionsId } = req;
+
+  if (process) {
+    await removeProcess(definitionsId);
+  }
+
+  /** @type {Ability} */
+  const userAbility = req.userAbility;
+
+  if (!userAbility.can('delete', toCaslResource('Process', process)))
+    return res.status(403).send('Forbidden.');
+
+  res.status(200).end();
+  return;
+});
+
+processRouter.get('/:definitionId/versions', isAllowed('view', 'Process'), async (req, res) => {
+  /** @type {Ability} */
+  const userAbility = req.userAbility;
+
+  if (!userAbility.can('view', toCaslResource('Process', req.process)))
+    return res.status(403).send('Forbidden.');
 
   res.status(200).send(req.process.versions);
 });
 
-processRouter.post(
-  '/:definitionId/versions',
-  isAllowed([PERMISSION_UPDATE, PERMISSION_MANAGE], 'Process'),
-  async (req, res) => {
-    const { definitionsId, body } = req;
+processRouter.post('/:definitionId/versions', isAllowed('update', 'Process'), async (req, res) => {
+  const { definitionsId, body } = req;
 
-    if (typeof body !== 'object' || !body.bpmn) {
-      res.status(400).send('Expected { bpmn: ... } as body!');
-      return;
-    }
+  /** @type {Ability} */
+  const userAbility = req.userAbility;
 
-    let { bpmn } = body;
+  if (!userAbility.can('update', toCaslResource('Process', req.process)))
+    return res.status(403).send('Forbidden.');
 
-    try {
-      await addProcessVersion(definitionsId, bpmn);
-    } catch (err) {
-      res.status(400).send(`Unable to add process version. Reason: ${err.message}`);
-      return;
-    }
+  if (typeof body !== 'object' || !body.bpmn) {
+    res.status(400).send('Expected { bpmn: ... } as body!');
+    return;
+  }
 
-    res.status(201).send();
-  },
-);
-
-processRouter.get('/:definitionId/versions/:version', async (req, res) => {
-  const { definitionsId } = req;
-  const { version } = req.params;
+  let { bpmn } = body;
 
   try {
-    res.status(200).send(await getProcessVersionBpmn(definitionsId, version));
+    await addProcessVersion(definitionsId, bpmn);
   } catch (err) {
-    res
-      .status(400)
-      .send(
-        `Unable to get version ${version} for the process (id: ${definitionsId})! Reason: ${err.message}.`,
-      );
+    res.status(400).send(`Unable to add process version. Reason: ${err.message}`);
+    return;
   }
+
+  res.status(201).send();
 });
+
+processRouter.get(
+  '/:definitionId/versions/:version',
+  isAllowed('view', 'Process'),
+  async (req, res) => {
+    const { definitionsId } = req;
+    const { version } = req.params;
+
+    /** @type {Ability} */
+    const userAbility = req.userAbility;
+
+    if (!userAbility.can('view', toCaslResource('Process', req.process)))
+      return res.status(403).send('Forbidden.');
+
+    try {
+      res.status(200).send(await getProcessVersionBpmn(definitionsId, version));
+    } catch (err) {
+      res
+        .status(400)
+        .send(
+          `Unable to get version ${version} for the process (id: ${definitionsId})! Reason: ${err.message}.`,
+        );
+    }
+  },
+);
 
 processRouter.use('/:definitionId/images', async (req, res, next) => {
   // early exit if the process doesn't exist
@@ -275,17 +244,33 @@ processRouter.use('/:definitionId/images', async (req, res, next) => {
   }
 });
 
-processRouter.get('/:definitionId/images', async (req, res) => {
+processRouter.get('/:definitionId/images', isAllowed('view', 'Process'), async (req, res) => {
+  /** @type {Ability} */
+  const userAbility = req.userAbility;
+
+  if (!userAbility.can('view', toCaslResource('Process', req.process)))
+    return res.status(403).send('Forbidden.');
+
   res.status(200).json(await getProcessImages(req.definitionsId));
 });
 
-processRouter.get('/:definitionId/images/:imageFileName', async (req, res) => {
-  const { definitionId, imageFileName } = req.params;
+processRouter.get(
+  '/:definitionId/images/:imageFileName',
+  isAllowed('view', 'Process'),
+  async (req, res) => {
+    const { definitionId, imageFileName } = req.params;
 
-  const image = await getProcessImage(definitionId, imageFileName);
-  res.set({ 'Content-Type': 'image/png image/svg+xml image/jpeg' });
-  res.status(200).send(image);
-});
+    /** @type {Ability} */
+    const userAbility = req.userAbility;
+
+    if (!userAbility.can('view', toCaslResource('Process', req.process)))
+      return res.status(403).send('Forbidden.');
+
+    const image = await getProcessImage(definitionId, imageFileName);
+    res.set({ 'Content-Type': 'image/png image/svg+xml image/jpeg' });
+    res.status(200).send(image);
+  },
+);
 
 processRouter.use('/:definitionId/user-tasks', async (req, res, next) => {
   // early exit if the process doesn't exist
@@ -296,8 +281,14 @@ processRouter.use('/:definitionId/user-tasks', async (req, res, next) => {
   }
 });
 
-processRouter.get('/:definitionId/user-tasks', async (req, res) => {
+processRouter.get('/:definitionId/user-tasks', isAllowed('view', 'Process'), async (req, res) => {
   let { withHtml } = req.query;
+
+  /** @type {Ability} */
+  const userAbility = req.userAbility;
+
+  if (!userAbility.can('view', toCaslResource('Process', req.process)))
+    return res.status(403).send('Forbidden.');
 
   if (withHtml === 'true') {
     res.status(200).json(await getProcessUserTasksHtml(req.definitionsId));
@@ -315,22 +306,38 @@ processRouter.use('/:definitionId/user-tasks/:userTaskFileName', async (req, res
   next();
 });
 
-processRouter.get('/:definitionId/user-tasks/:userTaskFileName', async (req, res) => {
-  const { userTaskFileName } = req;
+processRouter.get(
+  '/:definitionId/user-tasks/:userTaskFileName',
+  isAllowed('view', 'Process'),
+  async (req, res) => {
+    const { userTaskFileName } = req;
 
-  if (req.userTaskHtml) {
-    res.status(200).send(req.userTaskHtml);
-  } else {
-    res.status(404).send(`Found no html for user task filename ${userTaskFileName}`);
-  }
-});
+    /** @type {Ability} */
+    const userAbility = req.userAbility;
+
+    if (!userAbility.can('view', toCaslResource('Process', req.process)))
+      return res.status(403).send('Forbidden.');
+
+    if (req.userTaskHtml) {
+      res.status(200).send(req.userTaskHtml);
+    } else {
+      res.status(404).send(`Found no html for user task filename ${userTaskFileName}`);
+    }
+  },
+);
 
 processRouter.put(
   '/:definitionId/user-tasks/:userTaskFileName',
-  isAllowed([PERMISSION_UPDATE, PERMISSION_MANAGE], 'Process'),
+  isAllowed('update', 'Process'),
   async (req, res) => {
     const { definitionsId, userTaskFileName, userTaskHtml } = req;
     const { body } = req;
+
+    /** @type {Ability} */
+    const userAbility = req.userAbility;
+
+    if (!userAbility.can('update', toCaslResource('Process', req.process)))
+      return res.status(403).send('Forbidden.');
 
     await saveProcessUserTask(definitionsId, userTaskFileName, body);
 
@@ -346,9 +353,15 @@ processRouter.put(
 
 processRouter.delete(
   '/:definitionId/user-tasks/:userTaskFileName',
-  isAllowed([PERMISSION_UPDATE, PERMISSION_MANAGE], 'Process'),
+  isAllowed('update', 'Process'),
   async (req, res) => {
     const { definitionsId, userTaskFileName, userTaskHtml } = req;
+
+    /** @type {Ability} */
+    const userAbility = req.userAbility;
+
+    if (!userAbility.can('update', toCaslResource('Process', req.process)))
+      return res.status(403).send('Forbidden.');
 
     if (userTaskHtml) {
       await deleteProcessUserTask(definitionsId, userTaskFileName);
